pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "frpdeck"
version = "0.1.0"
description = "Structured FRP deployment and maintenance CLI"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "GitHub Copilot" }]
dependencies = [
  "jinja2>=3.1,<4",
  "pydantic>=2.7,<3",
  "PyYAML>=6.0,<7",
  "typer>=0.12,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
]

[project.scripts]
frpdeck = "frpdeck.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
frpdeck = ["templates/**/*.j2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
